Fixed test function
