+++ NinthHomework09.11.21/HashTable/hashTable.c
@@ -214,3 +214,24 @@ void printHashTable(HashTable* hashTable)
         }
     }
 }
+
+HashTable* createHashTableFromFile(char name[])
+{
+    FILE* file = fopen(name, "r");
+    HashTable* hashTable = createHashTable(1553);
+    while (!feof(file))
+    {
+        char* buffer = calloc(100, sizeof(char));
+        const int readBytes = fscanf(file, "%s", buffer);
+        if (readBytes < 0)
+        {
+            break;
+        }
+        addToHashTable(hashTable, buffer, 1);
+    }
+    recalculateTableStatistics(hashTable);
+    redoHashTable(&hashTable);
+    recalculateTableStatistics(hashTable);
+    fclose(file);
+    return hashTable;
+}
